# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/signup":{"post":{"operationId":"signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/login":{"post":{"operationId":"login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/google":{"get":{"operationId":"googleOAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/google/callback":{"get":{"operationId":"googleOAuthCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/logout":{"get":{"operationId":"logout","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/verify-session":{"get":{"operationId":"verifySession","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/booking":{"get":{"operationId":"getAllUserBookings","parameters":[{"name":"userId","required":true,"in":"query","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["booking"],"security":[{"cookie":[]}]},"post":{"operationId":"createBooking","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookingDto"}}}},"responses":{"201":{"description":""}},"tags":["booking"],"security":[{"cookie":[]},{"cookie":[]}]}},"/api/booking/{bookingId}":{"get":{"operationId":"getBooking","parameters":[{"name":"bookingId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["booking"],"security":[{"cookie":[]},{"cookie":[]}]},"patch":{"operationId":"editBooking","parameters":[{"name":"bookingId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookingDto"}}}},"responses":{"200":{"description":""}},"tags":["booking"],"security":[{"cookie":[]},{"cookie":[]}]},"delete":{"operationId":"deleteBooking","parameters":[{"name":"bookingId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["booking"],"security":[{"cookie":[]},{"cookie":[]}]}},"/api/users":{"get":{"operationId":"getAllUsers","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"cookie":[]}]}},"/api/users/{userId}":{"get":{"operationId":"getUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"cookie":[]}]},"patch":{"operationId":"editUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"],"security":[{"cookie":[]}]},"delete":{"operationId":"deleteUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"cookie":[]}]}},"/api/users/{userId}/password":{"patch":{"operationId":"changePassword","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["users"],"security":[{"cookie":[]}]}},"/api/users/{userId}/role":{"patch":{"operationId":"changeRole","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["users"],"security":[{"cookie":[]}]}},"/api/users/{userId}/active":{"patch":{"operationId":"changeActive","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeActiveDto"}}}},"responses":{"200":{"description":""}},"tags":["users"],"security":[{"cookie":[]}]}},"/api/users/{userId}/avatar":{"post":{"operationId":"uploadProfilePicture","parameters":[],"requestBody":{"required":true,"description":"Avatar","content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/"},"examples":{"file":{"value":"multipart/form-data"}}}}},"responses":{"201":{"description":""}},"tags":["users"],"security":[{"cookie":[]}]}},"/api/hotels":{"get":{"operationId":"getAllHotels","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hotels"]},"post":{"operationId":"createHotel","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHotelDto"}}}},"responses":{"201":{"description":""}},"tags":["hotels"],"security":[{"cookie":[]}]}},"/api/hotels/{hotelId}":{"get":{"operationId":"getHotel","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hotels"]},"patch":{"operationId":"editHotel","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHotelDto"}}}},"responses":{"200":{"description":""}},"tags":["hotels"],"security":[{"cookie":[]}]},"delete":{"operationId":"deleteHotel","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hotels"],"security":[{"cookie":[]}]}},"/api/hotels/{hotelId}/bookings":{"get":{"operationId":"getAllHotelBookings","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hotels"],"security":[{"cookie":[]}]}},"/api/hotels/{hotelId}/reviews":{"get":{"operationId":"getAllReviewsForHotel","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hotels"]},"post":{"operationId":"createReview","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":""}},"tags":["hotels"],"security":[{"cookie":[]}]}},"/api/hotels/{hotelId}/reviews/{reviewId}":{"patch":{"operationId":"editReview","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}},{"name":"reviewId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"200":{"description":""}},"tags":["hotels"],"security":[{"cookie":[]}]},"delete":{"operationId":"deleteReview","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}},{"name":"reviewId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hotels"],"security":[{"cookie":[]}]}},"/api/hotels/{hotelId}/categories":{"get":{"operationId":"getAllCategoriesForHotel","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hotels"]},"post":{"operationId":"addHotelCategory","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["hotels"],"security":[{"cookie":[]}]}},"/api/hotels/{hotelId}/categories/{categoryId}":{"delete":{"operationId":"deleteHotelCategory","parameters":[{"name":"hotelId","required":true,"in":"path","schema":{"type":"string"}},{"name":"categoryId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hotels"],"security":[{"cookie":[]}]}},"/api/categories":{"get":{"operationId":"getAllCategories","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["categories"]},"post":{"operationId":"createCategory","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["categories"],"security":[{"cookie":[]}]}},"/api/categories/{categoryId}":{"get":{"operationId":"getCategory","parameters":[{"name":"categoryId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["categories"]},"patch":{"operationId":"editCategory","parameters":[{"name":"categoryId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["categories"],"security":[{"cookie":[]}]},"delete":{"operationId":"deleteCategory","parameters":[{"name":"categoryId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["categories"],"security":[{"cookie":[]}]}},"/api/reward":{"get":{"operationId":"getAllHotelRewards","parameters":[{"name":"hotelId","required":true,"in":"query","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["reward"],"security":[{"cookie":[]}]},"post":{"operationId":"createReward","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRewardDto"}}}},"responses":{"201":{"description":""}},"tags":["reward"],"security":[{"cookie":[]},{"cookie":[]}]}},"/api/reward/{rewardId}":{"get":{"operationId":"getReward","parameters":[{"name":"rewardId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["reward"],"security":[{"cookie":[]},{"cookie":[]}]},"patch":{"operationId":"editReward","parameters":[{"name":"rewardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRewardDto"}}}},"responses":{"200":{"description":""}},"tags":["reward"],"security":[{"cookie":[]},{"cookie":[]}]},"delete":{"operationId":"deleteReward","parameters":[{"name":"rewardId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["reward"],"security":[{"cookie":[]},{"cookie":[]}]}}},"info":{"title":"BonAway API","description":"BonAway API description","version":"1.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"cookie":{"type":"apiKey","in":"cookie","name":"connect.sid"}},"schemas":{"SignupDto":{"type":"object","properties":{"userName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateBookingDto":{"type":"object","properties":{"description":{"type":"string"},"price":{"type":"number"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"hotelId":{"type":"number"},"userId":{"type":"number"}},"required":["description","hotelId","userId"]},"EditUserDto":{"type":"object","properties":{"userName":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}}},"ChangePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"}},"required":["oldPassword","newPassword"]},"ChangeRoleDto":{"type":"object","properties":{"role":{"type":"object"}},"required":["role"]},"ChangeActiveDto":{"type":"object","properties":{"active":{"type":"boolean"}},"required":["active"]},"CreateHotelDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"address":{"type":"string"},"description":{"type":"string"}},"required":["name","email","phoneNumber","address"]},"CreateReviewDto":{"type":"object","properties":{"text":{"type":"string"},"rating":{"type":"number"}},"required":["text","rating"]},"AddCategoryDto":{"type":"object","properties":{"categoryId":{"type":"number"}},"required":["categoryId"]},"CreateCategoryDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["name"]},"CreateRewardDto":{"type":"object","properties":{"description":{"type":"string"},"price":{"type":"number"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"hotelId":{"type":"number"},"userId":{"type":"number"}},"required":["description","hotelId","userId"]}}}}